chatbot-fastapi-nextjs/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.121.3",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "huggingface-hub>=1.1.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
